+++ lib/config.js
@@ -18,5 +18,7 @@ module.exports = {
 
     ROLLBAR_TOKEN: process.env.ROLLBAR_TOKEN, // should be undefined in dev
 
-    FB_VALIDATION_TOKEN: process.env.FB_VALIDATION_TOKEN || 'wtb_token'
+    FB_VALIDATION_TOKEN: process.env.FB_VALIDATION_TOKEN || 'wtb_token',
+    FB_PAGE_ACCESS_TOKEN: process.env.FB_PAGE_ACCESS_TOKEN,
+    FB_APP_SECRET: process.env.FB_APP_SECRET
 }
