+++ lib/index.js
@@ -287,11 +287,43 @@ function serviceExceptions() {
     }
     return noService;
 }
+
+/*
+ * Verify that the callback came from Facebook. Using the App Secret from
+ * the App Dashboard, we can verify the signature that is sent with each
+ * callback in the x-hub-signature field, located in the header.
+ *
+ * https://developers.facebook.com/docs/graph-api/webhooks#setup
+ *
+ */
+function verifyFBRequestSignature(req, res, buf) {
+    var signature = req.headers["x-hub-signature"];
+
+    if (!signature) {
+        // For testing, let's log an error. In production, you should throw an
+        // error.
+        throw new Error("Couldn't validate the signature.");
+    } else {
+        var elements = signature.split('=');
+        var method = elements[0];
+        var signatureHash = elements[1];
+
+        var expectedHash = crypto.createHmac('sha1', FB_APP_SECRET)
+            .update(buf)
+            .digest('hex');
+
+        if (signatureHash != expectedHash) {
+            throw new Error("Couldn't validate the request signature.");
+        }
+    }
+}
+
 module.exports.getStopsFromAddress = getStopsFromAddress;
 module.exports.getStopFromStopNumber = getStopFromStopNumber;
 module.exports.serviceExceptions = serviceExceptions;
 module.exports.findNearestStops = findNearestStops;
 module.exports.processFeedback = processFeedback;
+module.exports.verifyFBRequestSignature = verifyFBRequestSignature;
 
 // export this for unit testing purposes
 module.exports.getGeocodedAddress = getGeocodedAddress
