+++ routes/index.js
@@ -116,37 +116,6 @@ function getRoutes(req, res, next){
     }
 }
 
-/*
- * Verify that the callback came from Facebook. Using the App Secret from
- * the App Dashboard, we can verify the signature that is sent with each
- * callback in the x-hub-signature field, located in the header.
- *
- * https://developers.facebook.com/docs/graph-api/webhooks#setup
- *
- */
-function verifyRequestSignature(req, res, buf) {
-    var signature = req.headers["x-hub-signature"];
-
-    if (!signature) {
-        // For testing, let's log an error. In production, you should throw an
-        // error.
-        throw new Error("Couldn't validate the signature.");
-    } else {
-        var elements = signature.split('=');
-        var method = elements[0];
-        var signatureHash = elements[1];
-
-        var expectedHash = crypto.createHmac('sha1', APP_SECRET)
-            .update(buf)
-            .digest('hex');
-
-        if (signatureHash != expectedHash) {
-            throw new Error("Couldn't validate the request signature.");
-        }
-    }
-}
-
-
 /* GET HOME PAGE */
 router.get('/', function(req, res, next) {
         res.render('index');
@@ -235,7 +204,7 @@ function sendFBMessage(recipientId, messageText) {
 
     request({
         uri: 'https://graph.facebook.com/v2.6/me/messages',
-        qs: { access_token: PAGE_ACCESS_TOKEN },
+        qs: { access_token: FB_PAGE_ACCESS_TOKEN },
         method: 'POST',
         json: messageData
 
@@ -426,3 +395,4 @@ router.get('*', function(req, res){
 });
 
 module.exports = router;
+module.exports.verifyRequestSignature = verifyRequestSignature;
